# src/app/Http/Resources/CakeSolicitationResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Resources\Json\JsonResource;
use OpenApi\Attributes as OA;

/**
 * Class CakeSolicitationResource.
 *
 * @author ErickJMenezes <erickmenezes.dev@gmail.com>
 * @mixin \App\Models\CakeSolicitation
 */
#[OA\Schema(
    schema: 'CakeSolicitationResource',
    title: 'CakeSolicitation',
    description: 'Cake Solicitation Model',
    properties: [
        new OA\Property(
            property: 'email',
            description: 'Cake Solicitation Email',
            type: 'string',
            example: 'foo@example.com',
        ),
        new OA\Property(
            property: 'created_at',
            title: 'created_at',
            type: 'string',
            format: 'date-time-tz',
            example: '2020-01-01T00:00:00+00:00',
        ),
        new OA\Property(
            property: 'updated_at',
            title: 'updated_at',
            type: 'string',
            format: 'date-time-tz',
            example: '2020-01-01T00:00:00+00:00',
        ),
    ],
    type: 'object'
)]
class CakeSolicitationResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @param  \Illuminate\Http\Request  $request
     */
    public function toArray($request): array
    {
        return [
            'id' => $this->id,
            'email' => $this->email,
            'created_at' => $this->created_at,
            'updated_at' => $this->updated_at,
        ];
    }
}